# src/main/resources/swagger.yml
swagger: '2.0'
info:
  title: SpongeblogSP API
  description: Spongeblog blogging API
  version: "1.0.0"
# array of all schemes that your API supportsf
schemes:
  - https
# will be prefixed to all paths
basePath: /api/v1
consumes:
  - application/json
produces:
  - application/json
parameters:
  PageNumber:
    name: page
    in: query
    description: Page number from that on entities are returned.
    required: false
    type: integer
  ItemsPerPage:
    name: size
    in: query
    description: Entities per page. 
    required: false
    type: integer
...
  FromDate:
    name: from
    in: query
    description: Get items from timestamp.
    required: false
    type: integer
    format: int64
  ToDate:
    name: to
    in: query
    description: Get items until timestamp.
    required: false
    type: integer
    format: int64
  ImageKey:
    name: image-key
    in: path
    description: Key of the Image.
    required: true
    type: string
paths:
...
  /api/v1/images:
    post:
      summary: Uploads image file
      description: Uploads image file and returns their link.
      tags:
        - Images
      consumes:
        - multipart/form-data
      parameters:
        - name: file
          in: formData
          description: The uploaded image data
          required: true
          type: file
      responses:
        200:
          description: Image Link.
          schema:
            $ref: '#/definitions/ImageLink'
        400:
          description: Bad request (validation failed).
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized (need to log in / get token).
        403:
          description: Forbidden (no rights to access resource).
        404:
          description: Entity not found.
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request results in a conflict.
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error.
    get:
      summary: Get all Images.
      description: Returns a paginated list of all Images.
      tags:
        - Images
      parameters:
         - $ref: '#/parameters/PageNumber'
         - $ref: '#/parameters/ItemsPerPage'
         - $ref: '#/parameters/FromDate'
         - $ref: '#/parameters/ToDate'
      responses:
        200:
          description: Returns Image.
          schema:
            type: file
        400:
          description: Bad request (validation failed).
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized (need to log in / get token).
        403:
          description: Forbidden (no rights to access resource).
        404:
          description: Entity not found.
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request results in a conflict.
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error.
  /api/v1/images/{image-key}:
    get:
      summary: Load Image.
      description: Returns an Image file by key.
      tags:
        - Images
      parameters:
        - $ref: '#/parameters/ImageKey'
      responses:
        200:
          description: Returns Image.
          schema:
            type: file
        400:
          description: Bad request (validation failed).
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized (need to log in / get token).
        403:
          description: Forbidden (no rights to access resource).
        404:
          description: Entity not found.
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request results in a conflict.
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error.
    delete:
      summary: Removes Image.
      description: Removes Image from DB and storage.
      tags:
        - Images
      parameters:
        - $ref: '#/parameters/ImageKey'
      responses:
        200:
          description: Image deleted.
        400:
          description: Bad request (validation failed).
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized (need to log in / get token).
        403:
          description: Forbidden (no rights to access resource).
        404:
          description: Entity not found.
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request results in a conflict.
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error.
  /api/v1/images/link/{image-key}:
    get:
      summary: Load Image.
      description: Returns the link to the Image by key.
      tags:
        - Images
      parameters:
        - $ref: '#/parameters/ImageKey'
      responses:
        200:
          description: Returns Image link.
          schema:
            $ref: '#/definitions/ImageLink'
        400:
          description: Bad request (validation failed).
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized (need to log in / get token).
        403:
          description: Forbidden (no rights to access resource).
        404:
          description: Entity not found.
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request results in a conflict.
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error.
definitions:
  ImageLink:
    type: object
    properties:
      key:
        type: string
        description: Key of the Image (id + '_' +filename). 
      link:
        type: string
        description: URL of the Image.
...
  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      fields:
        type: string